# apidoc/dashboard-openapi.yaml
openapi: 3.0.0
info:
  title: Spotlight Dashboard REST API
  version: 1.0.0-20190117
servers:
- url: http://localhost:9000/dashapi/v1
...
security:
- bearer: []
paths:
...
  /receivers/:
    get:
      summary: Returns a list of receivers
      tags:
      - Receivers
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Receiver'
        400:
          description: Bad request
        401:
          description: Unauthorized
        403:
          description: Forbidden
    post:
      summary: Creates and returns a new receiver
      tags:
      - Receivers
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: "Receiver 1"
              required:
              - name
      responses:
        201:
          description: Successfully created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Receiver'
        400:
          description: Bad request
        401:
          description: Unauthorized
        403:
          description: Forbidden
  /receivers/{receiverId}:
    get:
      summary: Returns the receiver with receiverId.
      tags:
      - Receivers
      parameters:
      - name: receiverId
        in: path
        description: Receiver id
        required: true
        schema:
          type: integer
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Receiver'
    put:
      summary: Updates the receiver with receiverId.
      tags:
      - Receivers
      parameters:
      - name: receiverId
        in: path
        description: Receiver id
        required: true
        schema:
          type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: "Receiver 1"
              required:
              - name
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Receiver'
    delete:
      summary: Deletes the receiver with receiverId.
      description: Deletes the receiver with senderId. There should not be any process with the receiver to be deleted, otherwise it will return a Conflict error.
      tags:
      - Receivers
      parameters:
      - name: receiverId
        in: path
        description: Receiver id
        required: true
        schema:
          type: integer
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Receiver'
...
  /submissions/:
    get:
      summary: Returns a list of submissions.
      tags:
      - Submissions
      parameters:
      - name: sender
        in: query
        description: list of sender ids to filter for
        schema:
          type: array
          items:
            type: integer
        explode: false
      - name: receiver
        in: query
        description: list of receiver ids to filter for
        schema:
          type: array
          items:
            type: integer
        explode: false
      - name: processType
        in: query
        description: list of process type ids to filter for
        schema:
          type: array
          items:
            type: integer
        explode: false
      - name: from
        in: query
        description: only submissions with startTime after the beginning of this date will be returned
        schema:
          type: string
          format: date
      - name: to
        in: query
        description: only submissions with startTime before the beginning of this date will be returned
        schema:
          type: string
          format: date
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Submission'
...
  /schedules/:
    get:
      summary: Returns all the schedule definitions.
      tags:
      - Schedules
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ScheduleDefinition'
    post:
      summary: Creates and returns a new schedule definition.
      tags:
      - Schedules
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                processId:
                  type: integer
                  example: 1
                settings:
                  type: string
                  example: '{"dates": ["2019-02-28"]}'
                tolerance:
                  type: integer
                  example: 5
                scheduleChangeFlag:
                  type: boolean
                criticalDate:
                  type: string
                  format: date-time
                startTime:
                  type: string
                  format: date-time
                endTime:
                  type: string
                  format: date-time
                scheduleStartDate:
                  type: string
                  format: date-time
                scheduleEndDate:
                  type: string
                  format: date-time
                recurrencePattern:
                  type: string
                  example: 'C'
                recurrenceTime:
                  type: integer
                  example: 5
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ScheduleDefinition'
  /scheduled-submissions/:
    get:
      summary: Returns all the scheduled submissions from a date to a date.
      tags:
      - Schedules
      parameters:
      - name: from
        in: query
        description: from date
        required: true
        schema:
          type: string
          format: date
      - name: to
        in: query
        description: to date
        required: true
        schema:
          type: string
          format: date
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScheduledSubmission'
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
  schemas:
    ScheduledSubmission:
      type: object
      properties:
        id:
          type: integer
          example: 1
        scheduleDefId:
          type: integer
          example: 1
        submission:
          $ref: '#/components/schemas/Submission'
          nullable: true
        processId:
          type: integer
          example: 1
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        tolerance:
          type: integer
          example: 15
        notifications:
          type: integer
          example: 0
    ScheduleDefinition:
      type: object
      properties:
        id:
          type: integer
          example: 1
        process:
          $ref: '#/components/schemas/Process'
        processWorkflowId:
          type: integer
          example: 1
        settings:
          type: string
          example: '{"dates": ["2019-02-28"]}'
        tolerance:
          type: integer
          example: 5
        scheduleChangeFlag:
          type: boolean
        criticalDate:
          type: string
          format: date-time
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        scheduleStartDate:
          type: string
          format: date-time
        scheduleEndDate:
          type: string
          format: date-time
        recurrencePattern:
          type: string
          example: 'C'
        recurrenceTime:
          type: integer
          example: 5
...
    Submission:
      type: object
      properties:
        id:
          type: integer
          example: 1
        process:
          $ref: '#/components/schemas/Process'
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
          nullable: true
        updateTime:
          type: string
          format: date-time
        systemStartTime:
          type: string
          format: date-time
        systemEndTime:
          type: string
          format: date-time
          nullable: true
        systemUpdateTime:
          type: string
          format: date-time
        records:
          type: integer
          example: 10
        warnings:
          type: integer
          example: 10
        errors:
          type: integer
          example: 10
...
    ProcessType:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
    Receiver:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
    Role:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
    Sender:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
          example: "User 1"
        sso:
          type: integer
          example: 999999999
        roles:
          type: array
          items:
            $ref: '#/components/schemas/Role'
    Process:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
          example: "Process 1"
        sender:
          $ref: '#/components/schemas/Sender'
        receiver:
          $ref: '#/components/schemas/Receiver'
        processType:
          $ref: '#/components/schemas/ProcessType'
        critical:
          type: boolean
        supersession:
          type: boolean
        processParent:
          type: object
          nullable: true
        processLevel:
          type: integer
        isParent:
          type: boolean
        functionalOwner:
          $ref: '#/components/schemas/User'
        technicalOwner:
          $ref: '#/components/schemas/User'